+++ src/Fields.cpp
@@ -1,8 +1,9 @@
 #include <Geode/Geode.hpp>
-#include "../include/Fields.h"
-#include "../include/CCObject.h"
+#include "Fields.hpp"
+#include "ModifyCCObject.hpp"
 
 using namespace geode::prelude;
+using namespace alpha::utils;
 
 ObjectFieldContainer::~ObjectFieldContainer() {
     for (auto i = 0u; i < m_containedFields.size(); i++) {
@@ -27,7 +28,7 @@ void* ObjectFieldContainer::setField(size_t index, size_t size, std::function<vo
     return m_containedFields.at(index);
 }
 
-ObjectFieldContainer* ObjectFieldContainer::from(FieldCCObject* object, char const* forClass) {
+ObjectFieldContainer* ObjectFieldContainer::from(ModifyCCObject<CCObject>* object, char const* forClass) {
     return object->getFieldContainer(forClass);
 }
 
@@ -39,12 +40,10 @@ ObjectMetadata::~ObjectMetadata() {
     }
 }
 
-ObjectMetadata* ObjectMetadata::set(FieldCCObject* target) {
+ObjectMetadata* ObjectMetadata::set(ModifyCCObject<CCObject>* target) {
     if (!target) return nullptr;
 
     auto old = target->getUserObject();
-    // faster than dynamic_cast, technically can
-    // but extremely unlikely to fail
     if (old && old->getTag() == METADATA_TAG) {
         return static_cast<ObjectMetadata*>(old);
     }
@@ -52,13 +51,11 @@ ObjectMetadata* ObjectMetadata::set(FieldCCObject* target) {
     meta->autorelease();
     meta->setTag(METADATA_TAG);
 
-    // set user object
     target->setUserObject(meta);
     meta->retain();
 
     if (old) {
         meta->m_userObjects.insert({ "", old });
-        // the old user object is now managed by Ref
         old->release();
     }
     return meta;
@@ -74,4 +71,3 @@ ObjectFieldContainer* ObjectMetadata::getFieldContainer(char const* forClass) {
 
     return container;
 }
-